+++ examples/one_dot_one/src/generatedNoCheck/objects/index.ts
@@ -1,3 +1,0 @@
-export * from './ObjectTypeWithAllPropertyTypes';
-export * from './Person';
-export * from './Todo';