+++ examples/one_dot_one/src/generatedNoCheck/ontologyObjects.ts
@@ -1,8 +1,0 @@
-import { BaseObjectSet } from '@osdk/legacy-client';
-import { ObjectTypeWithAllPropertyTypes, Person, Todo } from './objects';
-
-export interface Objects {
-  Todo: BaseObjectSet<Todo>;
-  Person: BaseObjectSet<Person>;
-  ObjectTypeWithAllPropertyTypes: BaseObjectSet<ObjectTypeWithAllPropertyTypes>;
-}