+++ packages/generator/src/v1.1/generateClientSdkVersionOneDotOne.ts
@@ -23,6 +23,7 @@ import { generateFoundryClientFile } from "./generateFoundryClientFile";
 import { generateIndexFile } from "./generateIndexFile";
 import { generateMetadataFile } from "./generateMetadataFile";
 import { generateObjectsInterfaceFile } from "./generateObjectsInterfaceFile";
+import { generateOntologyIndexFile } from "./generateOntologyIndexFile";
 import { generatePerActionDataFiles } from "./generatePerActionDataFiles";
 import { generatePerObjectInterfaceAndDataFiles } from "./generatePerObjectInterfaceAndDataFiles";
 import { generatePerQueryDataFiles } from "./generatePerQueryDataFiles";
@@ -33,26 +34,31 @@ export async function generateClientSdkVersionOneDotOne(
   fs: MinimalFs,
   outDir: string,
 ) {
+  const objectsDir = path.join(outDir, "ontology", "objects");
+  const actionsDir = path.join(outDir, "ontology", "actions");
+  const queriesDir = path.join(outDir, "ontology", "queries");
+
   const sanitizedOntology = sanitizeMetadata(ontology);
   await generateFoundryClientFile(fs, outDir);
   await generateMetadataFile(sanitizedOntology, fs, outDir);
-  await generateObjectsInterfaceFile(sanitizedOntology, fs, outDir);
+  await generateOntologyIndexFile(fs, path.join(outDir, "ontology"));
+  await generateObjectsInterfaceFile(sanitizedOntology, fs, objectsDir);
   await generatePerObjectInterfaceAndDataFiles(
     sanitizedOntology,
     fs,
-    path.join(outDir, "objects"),
+    objectsDir,
   );
-  await generateActions(sanitizedOntology, fs, outDir);
+  await generateActions(sanitizedOntology, fs, actionsDir);
   await generatePerActionDataFiles(
     sanitizedOntology,
     fs,
-    path.join(outDir, "actions"),
+    actionsDir,
   );
-  await generateQueries(sanitizedOntology, fs, outDir);
+  await generateQueries(sanitizedOntology, fs, queriesDir);
   await generatePerQueryDataFiles(
     sanitizedOntology,
     fs,
-    path.join(outDir, "queries"),
+    queriesDir,
   );
   await generateIndexFile(fs, outDir);
 }