+++ packages/generator/src/v1.1/generateMetadataFile.ts
@@ -35,19 +35,25 @@ export async function generateMetadataFile(
     await formatTs(`
   import type { OntologyDefinition } from "@osdk/api";
   import type { Ontology as ClientOntology } from "@osdk/legacy-client";
-  import type { Objects } from "./ontologyObjects";
-  import type { Actions } from "./ontologyActions";
-  import type { Queries } from "./ontologyQueries";
+  import type { Objects } from "./ontology/objects/Objects";
+  import type { Actions } from "./ontology/actions/Actions";
+  import type { Queries } from "./ontology/queries/Queries";
   ${
-      objectNames.map((name) => `import {${name}} from "./objects/${name}";`)
+      objectNames.map((name) =>
+        `import {${name}} from "./ontology/objects/${name}";`
+      )
         .join("\n")
     }
   ${
-      actionNames.map((name) => `import {${name}} from "./actions/${name}";`)
+      actionNames.map((name) =>
+        `import {${name}} from "./ontology/actions/${name}";`
+      )
         .join("\n")
     }
   ${
-      queryNames.map(name => `import {${name}} from "./queries/${name}";`).join(
+      queryNames.map(name =>
+        `import {${name}} from "./ontology/queries/${name}";`
+      ).join(
         "\n",
       )
     }