+++ packages/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/api
 
+## 0.0.10
+
+### Patch Changes
+
+- 9167059: Fixing GeoJson import from namespace
+
+## 0.0.9
+
+### Patch Changes
+
+- 0860ae9: Adds type information for geopoint and geoshape
+
 ## 0.0.8
 
 ### Patch Changes