+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @osdk/cli
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [9167059]
+  - @osdk/api@0.0.10
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [0860ae9]
+  - @osdk/api@0.0.9
+
 ## 0.0.4
 
 ### Patch Changes