+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @osdk/generator
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [9167059]
+  - @osdk/api@0.0.10
+
+## 0.0.7
+
+### Patch Changes
+
+- 7fc9fdc: Adjusting package structure to match existing
+- Updated dependencies [0860ae9]
+  - @osdk/api@0.0.9
+
 ## 0.0.6
 
 ### Patch Changes